Services/django_app/requirements.txt
django
gunicorn
djangorestframework
djangorestframework-simplejwt
requests
python-dotenv
whitenoise
